changing color of the map 2

+++ src/player.c
@@ -6,7 +6,7 @@
 /*   By: yacinebentayeb <yacinebentayeb@student.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/26 00:20:05 by yacinebenta       #+#    #+#             */
-/*   Updated: 2022/07/31 21:12:46 by yacinebenta      ###   ########.fr       */
+/*   Updated: 2022/08/01 14:37:23 by yacinebenta      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -104,34 +104,44 @@ void	update_angle(int key, t_data *data)
 
 void	update_position(int key, t_data *data)
 {
-	int		speed;
+	long double		speed;
 	float	tmp_angle;
 
 	speed = (int)data->map->tile_size / 9;
+	// if (speed < 2)
+	// 	speed = 1.5;
 	if (key == KEY_W || key == KEY_W_L)
 	{
-		data->p->d_x = data->p->x + (int)((float)cos((double)data->p->angle) * speed);
-		data->p->d_y = data->p->y - (int)((float)sin((double)data->p->angle) * speed);
+		data->p->d_x = data->p->x +
+			(int)((float)cos((double)data->p->angle) * speed);
+		data->p->d_y = data->p->y -
+			(int)((float)sin((double)data->p->angle) * speed);
 	}
 	if (key == KEY_A || key == KEY_A_L)
 	{
 		tmp_angle = data->p->angle + 90. * (M_PI / 180.0);
-		data->p->d_x = data->p->x + (int)((float)cos((double)tmp_angle) * speed);
-		data->p->d_y = data->p->y - (int)((float)sin((double)tmp_angle) * speed);
+		data->p->d_x = data->p->x +
+			(int)((float)cos((double)tmp_angle) * speed);
+		data->p->d_y = data->p->y -
+			(int)((float)sin((double)tmp_angle) * speed);
 	}
 	if (key == KEY_S || key == KEY_S_L)
 	{
 		tmp_angle = data->p->angle - 90. * (M_PI / 180.0);
-		data->p->d_x = data->p->x + (int)((float)cos((double)tmp_angle) * speed);
-		data->p->d_y = data->p->y - (int)((float)sin((double)tmp_angle) * speed);
+		data->p->d_x = data->p->x +
+			(int)((float)cos((double)tmp_angle) * speed);
+		data->p->d_y = data->p->y -
+			(int)((float)sin((double)tmp_angle) * speed);
 	}
 	if (key == KEY_D || key == KEY_D_L)
 	{
 		tmp_angle = data->p->angle + 180. * (M_PI / 180.0);
-		data->p->d_x = data->p->x + (int)((float)cos((double)tmp_angle) * speed);
-		data->p->d_y = data->p->y - (int)((float)sin((double)tmp_angle) * speed);
+		data->p->d_x = data->p->x +
+			(int)((float)cos((double)tmp_angle) * speed);
+		data->p->d_y = data->p->y -(int)((float)sin((double)tmp_angle) * speed);
 	}
-	if (data->map->map2d[(int)floor(data->p->d_y / data->map->tile_size)][(int)floor(data->p->d_x / data->map->tile_size)] != '1')
+	if (data->map->map2d[(int)floor(data->p->d_y/ data->map->tile_size)]
+		[(int)floor(data->p->d_x / data->map->tile_size)] != '1')
 	{
 		if (data->map->map2d[(int)floor(data->p->d_y / data->map->tile_size)][(int)floor(data->p->d_x / data->map->tile_size)] == 'D')
 			data->map->map2d[(int)floor(data->p->d_y / data->map->tile_size)][(int)floor(data->p->d_x / data->map->tile_size)] = '0';
